Create FOCMotorSim
WPILib's DCMotorSim class works well for standard voltage control, but breaks if you try to run current control. This is a problem since we use TorqueCurrentFOC for the drive motors. FOCMotorSim uses a super simple simulation of a linear system (modeled as a flywheel) that coasts indefinitely. Modelling friction is left to the user.

- Tada