Remove supply current limiting from SwerveSetpointGenerator
Either my dynamics model in FOCMotorSim is wrong or there's jank in SwerveSetpointGenerator, either way the simulated motor dynamics couldn't follow the setpoint.

On top of this, supply current limits makes the motor dynamics harder to intuit, whilst not making much of a difference in terms of drawn supply current. The sharp peaks in current draw from not having a supply current limit is easily absorbed by internal capacitance in a battery and aside from that peak, supply current tends to be fairly reasonable.

- Tada